requirements.txt
pytube>=12.1.0
google-api-python-client>=2.0.0
pyannote.audio>=2.1.1
openai>=0.27.0
langchain>=0.0.267
whisper>=1.0.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
pydub>=0.25.1
python-dotenv>=1.0.0
langchain-community>=0.0.26
# Testing dependencies
pytest-vcr>=1.0.2
vcrpy>=4.2.1